+++ jklib/django/commands/__init__.py
@@ -1,5 +1,0 @@
-"""
-Provides classes, functions, and various utilities for creating Django commands
-    commands.py     Improvement over the main BaseCommand class
-    operations.py   Provides classes to perform the actual operations of a command
-"""