+++ jklib/django/db/__init__.py
@@ -1,1 +1,0 @@
-"""Package with utility functions that focuses on django models"""