+++ jklib/django/drf/__init__.py
@@ -1,1 +1,0 @@
-"""Package with utility functions for the django rest framework 3rd party package"""