+++ jklib/std/maths.py
@@ -1,22 +1,12 @@
 """Utility functions for maths and working with numbers in general"""
 
-
 # Built-in
 from decimal import Decimal
+from typing import Union
 
 
-# --------------------------------------------------------------------------------
-# > Functions
-# --------------------------------------------------------------------------------
-def precise_round(value, points: int):
-    """
-    Improved round() function using the Decimal class
-    :param value: The value the round
-    :type value: int or float
-    :param int points: The number of decimals you want
-    :return: The rounded value as a Decimal instance
-    :rtype: Decimal
-    """
+def precise_round(value: Union[int, float], points: int) -> Decimal:
+    """Improved round() function using the Decimal class"""
     rounded_value = round(value, points)
     precision = f".{'0'*(points-1)}1"
     return Decimal(rounded_value).quantize(Decimal(precision))