+++ jklib/std/threads.py
@@ -1,19 +1,11 @@
 """Utility functions and classes for threads"""
-
-
 # Built-in
 from threading import Thread
+from typing import List
 
 
-# --------------------------------------------------------------------------------
-# > Functions
-# --------------------------------------------------------------------------------
-def run_all_threads(threads):
-    """
-    Runs all the generated threads at once, and then joins them
-    :param threads: List of Thread instances we want to run
-    :type threads: list(Thread)
-    """
+def run_all_threads(threads: List[Thread]) -> None:
+    """Runs all the generated threads at once, and then joins them"""
     for thread in threads:
         thread.start()
     for thread in threads: